backend/requirements.txt
# requirements.txt
# EduVerify Backend Dependencies
# Python 3.10+ required

# ============================================
# Core Framework
# ============================================
fastapi
uvicorn
python-dotenv

# ============================================
# Database
# ============================================
sqlalchemy
asyncpg
alembic
psycopg2-binary

# ============================================
# Authentication & Security
# ============================================
python-jose
passlib[bcrypt]
python-multipart
cryptography
bcrypt

# ============================================
# Data Validation & Serialization
# ============================================
pydantic
pydantic-settings
email-validator

# ============================================
# Caching & Message Queue
# ============================================
redis
aioredis
celery

# ============================================
# HTTP Client & API Integration
# ============================================
httpx
aiohttp
requests

# ============================================
# AI & Computer Vision
# ============================================
# torch==2.1.0
# torchvision==0.16.0
# opencv-python==4.8.1.78
# numpy==1.24.3
# pillow==10.1.0

# YOLOv8-Face (Face Detection)
# ultralytics==8.0.221

# MediaPipe (Face Landmarks)
# mediapipe==0.10.7

# ArcFace (Face Recognition)
# insightface==0.7.3

# DINOv2 (Liveness Detection)
# transformers==4.36.0

# ============================================
# Data Processing
# ============================================
pandas
scikit-learn
scipy

# ============================================
# Logging & Monitoring
# ============================================
python-json-logger
prometheus-client
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
# ============================================
# File Processing
# ============================================
openpyxl
pandas
csvkit

# ============================================
# Task Scheduling
# ============================================
apscheduler
schedule

# ============================================
# Email & Notifications
# ============================================
# sendgrid==6.10.0
# fastapi-mail==1.4.1

# ============================================
# Image Processing
# ============================================
# imageio==2.33.1
# face-recognition==1.3.0
# facenet-pytorch==2.5.2

# ============================================
# Testing
# ============================================
# pytest==7.4.3
# pytest-asyncio==0.21.1
# pytest-cov==4.1.0
# httpx==0.25.1
# factory-boy==3.3.0
# faker==20.1.0

# ============================================
# Development Tools
# ============================================
# black==23.11.0
# isort==5.12.0
# flake8==6.1.0
# mypy==1.7.1
# pre-commit==3.5.0

# ============================================
# Documentation
# ============================================
# mkdocs==1.5.3
# mkdocstrings-python==1.7.5

# ============================================
# Deployment & DevOps
# ============================================
# docker==6.1.3
# gunicorn==21.2.0
# supervisor==4.2.5

# ============================================
# Utilities
# ============================================
# python-dateutil==2.8.2
# pytz==2023.3
# typing-extensions==4.8.0
# orjson==3.9.10
# ujson==5.8.0